% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstat.R
\name{calculate_mean}
\alias{calculate_mean}
\title{1.Calculate the mean of a numeric vector.}
\usage{
calculate_mean(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
The mean value of data if it is not empty; otherwise, an error message.
}
\description{
This function calculates the mean of a numeric vector. If the vector is empty,
it returns an error message.
}
