% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstat.R
\name{calculate_median}
\alias{calculate_median}
\title{2.Calculate the median of a numeric vector.}
\usage{
calculate_median(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
The median value of data if it is not empty; otherwise, an error message.
}
\description{
This function calculates the median of a numeric vector. If the vector is empty,
it returns an error message.
}
