% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstat.R
\name{calculate_mode}
\alias{calculate_mode}
\title{3.Calculate the mode(s) of a numeric vector.}
\usage{
calculate_mode(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
The mode(s) of data if it exists; otherwise, an error message or "No mode found".
}
\description{
This function calculates the mode(s) of a numeric vector. If the vector is empty,
it returns an error message. If there is no mode (all values are unique), it
returns "No mode found".
}
