% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstat.R
\name{calculate_standard_deviation}
\alias{calculate_standard_deviation}
\title{5.Calculate the standard deviation of a numeric vector.}
\usage{
calculate_standard_deviation(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The standard deviation of x.
}
\description{
5.Calculate the standard deviation of a numeric vector.
}
