% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstat.R
\name{calculate_variance}
\alias{calculate_variance}
\title{4.Calculate the variance of a numeric vector.}
\usage{
calculate_variance(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The variance of x.
}
\description{
4.Calculate the variance of a numeric vector.
}
